integer

+++ database/migrations/2024_12_13_164947_add_numero_tarjeta_to_adicionales_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('adicionales', function (Blueprint $table) {
-            $table->unsignedBigInteger('numero_tarjeta')->nullable()->after('cuenta');
+            $table->integer('numero_tarjeta')->nullable()->after('cuenta');
         });
     }
